feat: add utility functions for badge and button variants, update component exports
