+++ src/components/ui/badge.tsx
@@ -1,28 +1,7 @@
 import * as React from 'react';
-import { cva, type VariantProps } from 'class-variance-authority';
+import { type VariantProps } from 'class-variance-authority';
 import { cn } from '@/lib/utils';
-
-const badgeVariants = cva(
-  'inline-flex items-center rounded-full border px-2.5 py-0.5 text-xs font-semibold transition-colors focus:outline-none focus:ring-2 focus:ring-ring focus:ring-offset-2',
-  {
-    variants: {
-      variant: {
-        default: 'border-transparent bg-primary text-primary-foreground',
-        secondary: 'border-transparent bg-secondary text-secondary-foreground',
-        destructive: 'border-transparent bg-destructive text-destructive-foreground',
-        outline: 'text-foreground',
-        pending: 'border-transparent bg-yellow-100 text-yellow-800',
-        contacted: 'border-transparent bg-blue-100 text-blue-800',
-        confirmed: 'border-transparent bg-green-100 text-green-800',
-        delivered: 'border-transparent bg-gray-100 text-gray-800',
-        cancelled: 'border-transparent bg-red-100 text-red-800',
-      },
-    },
-    defaultVariants: {
-      variant: 'default',
-    },
-  }
-);
+import { badgeVariants } from './utils';
 
 export interface BadgeProps
   extends React.HTMLAttributes<HTMLDivElement>,
@@ -32,4 +11,4 @@ function Badge({ className, variant, ...props }: BadgeProps) {
   return <div className={cn(badgeVariants({ variant }), className)} {...props} />;
 }
 
-export { Badge, badgeVariants };
+export { Badge };
