+++ src/components/ui/index.ts
@@ -1,4 +1,4 @@
-export { Button, buttonVariants } from "./button";
+export { Button } from "./button";
 export { Input } from "./input";
 export { Textarea } from "./textarea";
 export {
@@ -9,9 +9,10 @@ export {
   CardDescription,
   CardContent,
 } from "./card";
-export { Badge, badgeVariants } from "./badge";
+export { Badge } from "./badge";
 export { Label } from "./label";
 export { Select } from "./select";
 export { Spinner, LoadingScreen, LoadingSpinner } from "./spinner";
 export { ConfirmDialogContainer } from "./confirm-dialog";
 export { ToastContainer } from "./toast-container";
+export { badgeVariants, buttonVariants } from "./utils";
