+++ src/components/ui/label.tsx
@@ -1,9 +1,7 @@
 import * as React from 'react';
 import { cn } from '@/lib/utils';
 
-interface LabelProps extends React.LabelHTMLAttributes<HTMLLabelElement> {}
-
-const Label = React.forwardRef<HTMLLabelElement, LabelProps>(({ className, ...props }, ref) => (
+const Label = React.forwardRef<HTMLLabelElement, React.LabelHTMLAttributes<HTMLLabelElement>>(({ className, ...props }, ref) => (
   <label
     ref={ref}
     className={cn(
